Travis specific test tweaks
Do not try to run main.js on Travis since it depends on a custom node
runtime that we don't build on Travis since that would require
30min test runs instead of 1min runs.

+++ scripts/test.sh
@@ -24,17 +24,21 @@ if ! scripts/generate-production-source-list.js generated/prod-sources; then
     fi
 fi
 
-# Sanity check the server by starting it, and then tell it to hangup.
-echo Starting server
-./main.js &
-main_pid="$!"
-# Spawn a fire-and-forget subprocess to stop the server from listening forever.
-(
-    sleep 1
-    kill -s INT "$main_pid" >& /dev/null
-) &
-hup_pid="$!"
-wait "$main_pid"  # Wait for the server to exit
-server_result="$?"
-kill "$hup_pid" >& /dev/null || true
-exit "$server_result"
+if [[ "$TRAVIS" != "true" ]]; then
+    # Sanity check the server by starting it, and then tell it to hangup.
+    echo Starting server
+    ./main.js &
+    main_pid="$!"
+    # Spawn a fire-and-forget subprocess to stop the server from listening forever.
+    (
+        sleep 1
+        kill -s INT "$main_pid" >& /dev/null
+    ) &
+    hup_pid="$!"
+    wait "$main_pid"  # Wait for the server to exit
+    server_result="$?"
+    kill "$hup_pid" >& /dev/null || true
+    if [[ "$server_result" != "0" ]]; then
+        exit "$server_result"
+    fi
+fi
